+++ common/mmodel/asset.go
@@ -5,34 +5,32 @@ import "time"
 // CreateAssetInput is a struct design to encapsulate request create payload data.
 //
 // swagger:model CreateAssetInput
-// @Description CreateAssetInput is a struct design to encapsulate request create payload data.
-// @Param name query string false "Name" e
-// @Param type query string false "Type"
-// @Param code query string false "Code"
-// @Param status query string false "Status"
-// @Param metadata query string false "Metadata"
+//
+//	@Description	CreateAssetInput is the input payload to create an asset.
 type CreateAssetInput struct {
 	Name     string         `json:"name" validate:"max=256" example:"Brazilian Real"`
 	Type     string         `json:"type" example:"currency"`
 	Code     string         `json:"code" validate:"required,max=100" example:"BRL"`
 	Status   Status         `json:"status"`
 	Metadata map[string]any `json:"metadata" validate:"dive,keys,keymax=100,endkeys,nonested,valuemax=2000"`
-} // @name CreateAssetInput
+} //	@name	CreateAssetInput
 
 // UpdateAssetInput is a struct design to encapsulate request update payload data.
 //
 // swagger:model UpdateAssetInput
-// @Description UpdateAssetInput is a struct design to encapsulate request update payload data.
+//
+//	@Description	UpdateAssetInput is the input payload to update an asset.
 type UpdateAssetInput struct {
 	Name     string         `json:"name" validate:"max=256" example:"Bitcoin"`
 	Status   Status         `json:"status"`
 	Metadata map[string]any `json:"metadata" validate:"dive,keys,keymax=100,endkeys,nonested,valuemax=2000"`
-} // @name UpdateAssetInput
+} //	@name	UpdateAssetInput
 
 // Asset is a struct designed to encapsulate payload data.
 //
 // swagger:model Asset
-// @Description Asset is a struct designed to encapsulate payload data.
+//
+//	@Description	Asset is a struct designed to store asset data.
 type Asset struct {
 	ID             string         `json:"id" example:"00000000-0000-0000-0000-000000000000"`
 	Name           string         `json:"name" example:"Brazilian Real"`
@@ -45,17 +43,15 @@ type Asset struct {
 	UpdatedAt      time.Time      `json:"updatedAt" example:"2021-01-01T00:00:00Z"`
 	DeletedAt      *time.Time     `json:"deletedAt" example:"2021-01-01T00:00:00Z"`
 	Metadata       map[string]any `json:"metadata,omitempty"`
-} // @name Asset
+} //	@name	Asset
 
 // Assets struct to return get all.
 //
 // swagger:model Assets
-// @Description Assets struct to return get all.
-// @Param items query string false "Items"
-// @Param page query string false "Page"
-// @Param limit query string false "Limit"
+//
+//	@Description	Assets is the struct designed to return a list of assets with pagination.
 type Assets struct {
 	Items []Asset `json:"items"`
 	Page  int     `json:"page" example:"1"`
 	Limit int     `json:"limit" example:"10"`
-} // @name Assets
+} //	@name	Assets