+++ components/ledger/api/docs.go
@@ -39,6 +39,12 @@ const docTemplate = `{
                         "description": "Metadata",
                         "name": "metadata",
                         "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -86,6 +92,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreateOrganizationInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -115,6 +127,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -139,6 +157,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -175,6 +199,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdateOrganizationInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -211,6 +241,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -265,6 +301,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreateLedgerInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -294,6 +336,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -325,6 +373,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -368,6 +422,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdateLedgerInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -410,6 +470,12 @@ const docTemplate = `{
                         "description": "Metadata",
                         "name": "metadata",
                         "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -471,6 +537,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreateAccountInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -514,6 +586,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -552,6 +630,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -602,6 +686,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdateAccountInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -644,6 +734,12 @@ const docTemplate = `{
                         "description": "Metadata",
                         "name": "metadata",
                         "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -705,6 +801,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreateAssetInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -748,6 +850,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -786,6 +894,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -836,6 +950,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdateAssetInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -878,6 +998,12 @@ const docTemplate = `{
                         "description": "Metadata query",
                         "name": "metadata",
                         "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -939,6 +1065,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreatePortfolioInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -982,6 +1114,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1020,6 +1158,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1063,6 +1207,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdatePortfolioInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1112,6 +1262,12 @@ const docTemplate = `{
                         "description": "Metadata",
                         "name": "metadata",
                         "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1180,6 +1336,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreateAccountInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1230,6 +1392,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1268,6 +1436,12 @@ const docTemplate = `{
                         "name": "portfolio_id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1325,6 +1499,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdateAccountInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1367,6 +1547,12 @@ const docTemplate = `{
                         "description": "Metadata",
                         "name": "metadata",
                         "in": "query"
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1428,6 +1614,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/CreateProductInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1471,6 +1663,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1509,6 +1707,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1559,6 +1763,12 @@ const docTemplate = `{
                         "schema": {
                             "$ref": "#/definitions/UpdateProductInput"
                         }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Request ID",
+                        "name": "Midaz-Id",
+                        "in": "header"
                     }
                 ],
                 "responses": {
@@ -1574,7 +1784,7 @@ const docTemplate = `{
     },
     "definitions": {
         "Account": {
-            "description": "Account is a struct designed to encapsulate response payload data.",
+            "description": "Account is a struct designed to store account data.",
             "type": "object",
             "properties": {
                 "alias": {
@@ -1654,7 +1864,7 @@ const docTemplate = `{
             }
         },
         "Address": {
-            "description": "Address structure for marshaling/unmarshalling JSON.",
+            "description": "Address is a struct designed to store the address data of an organization.",
             "type": "object",
             "properties": {
                 "city": {
@@ -1685,7 +1895,7 @@ const docTemplate = `{
             }
         },
         "Asset": {
-            "description": "Asset is a struct designed to encapsulate payload data.",
+            "description": "Asset is a struct designed to store asset data.",
             "type": "object",
             "properties": {
                 "code": {
@@ -1734,7 +1944,7 @@ const docTemplate = `{
             }
         },
         "Balance": {
-            "description": "Balance structure for marshaling/unmarshalling JSON.",
+            "description": "Balance is the struct designed to represent the account balance.",
             "type": "object",
             "properties": {
                 "available": {
@@ -1752,7 +1962,7 @@ const docTemplate = `{
             }
         },
         "CreateAccountInput": {
-            "description": "CreateAccountInput is a struct design to encapsulate request create payload data.",
+            "description": "CreateAccountInput is the input payload to create an account.",
             "type": "object",
             "required": [
                 "assetCode",
@@ -1813,7 +2023,7 @@ const docTemplate = `{
             }
         },
         "CreateAssetInput": {
-            "description": "CreateAssetInput is a struct design to encapsulate request create payload data.",
+            "description": "CreateAssetInput is the input payload to create an asset.",
             "type": "object",
             "required": [
                 "code"
@@ -1843,7 +2053,7 @@ const docTemplate = `{
             }
         },
         "CreateLedgerInput": {
-            "description": "CreateLedgerInput is a struct design to encapsulate request create payload data for ledger.",
+            "description": "CreateLedgerInput is the input payload to create a ledger.",
             "type": "object",
             "required": [
                 "name"
@@ -1864,7 +2074,7 @@ const docTemplate = `{
             }
         },
         "CreateOrganizationInput": {
-            "description": "CreateOrganizationInput is a struct design to encapsulate request create payload data for organization.",
+            "description": "CreateOrganizationInput is the input payload to create an organization.",
             "type": "object",
             "required": [
                 "legalDocument",
@@ -1903,7 +2113,7 @@ const docTemplate = `{
             }
         },
         "CreatePortfolioInput": {
-            "description": "CreatePortfolioInput is a struct design to encapsulate request create payload data.",
+            "description": "CreatePortfolioInput is the input payload to create a portfolio.",
             "type": "object",
             "required": [
                 "entityId",
@@ -1930,7 +2140,7 @@ const docTemplate = `{
             }
         },
         "CreateProductInput": {
-            "description": "CreateProductInput is a struct design to encapsulate request create payload data.",
+            "description": "CreateProductInput is the input payload to create a product.",
             "type": "object",
             "required": [
                 "name"
@@ -1951,7 +2161,7 @@ const docTemplate = `{
             }
         },
         "Ledger": {
-            "description": "Ledger is a struct designed to encapsulate payload data.",
+            "description": "Ledger is a struct designed to store ledger data.",
             "type": "object",
             "properties": {
                 "createdAt": {
@@ -1988,7 +2198,7 @@ const docTemplate = `{
             }
         },
         "Organization": {
-            "description": "Organization is a struct designed to encapsulate response payload data.",
+            "description": "Organization is a struct designed to store organization data.",
             "type": "object",
             "properties": {
                 "address": {
@@ -2036,7 +2246,7 @@ const docTemplate = `{
             }
         },
         "Pagination": {
-            "description": "Pagination is a struct designed to encapsulate pagination response payload data.",
+            "description": "Pagination is the struct designed to store the pagination data of an entity list.",
             "type": "object",
             "properties": {
                 "items": {},
@@ -2051,7 +2261,7 @@ const docTemplate = `{
             }
         },
         "Portfolio": {
-            "description": "Portfolio is a struct designed to encapsulate request update payload data.",
+            "description": "Portfolio is a struct designed to store portfolio data.",
             "type": "object",
             "properties": {
                 "createdAt": {
@@ -2096,7 +2306,7 @@ const docTemplate = `{
             }
         },
         "Product": {
-            "description": "Product is a struct designed to encapsulate payload data.",
+            "description": "Product is a struct designed to store product data.",
             "type": "object",
             "properties": {
                 "createdAt": {
@@ -2137,7 +2347,7 @@ const docTemplate = `{
             }
         },
         "Status": {
-            "description": "Status structure for marshaling/unmarshalling JSON.",
+            "description": "Status is the struct designed to store the status data of an entity.",
             "type": "object",
             "properties": {
                 "code": {
@@ -2153,7 +2363,7 @@ const docTemplate = `{
             }
         },
         "UpdateAccountInput": {
-            "description": "UpdateAccountInput is a struct design to encapsulate request update payload data.",
+            "description": "UpdateAccountInput is the input payload to update an account.",
             "type": "object",
             "properties": {
                 "alias": {
@@ -2188,7 +2398,7 @@ const docTemplate = `{
             }
         },
         "UpdateAssetInput": {
-            "description": "UpdateAssetInput is a struct design to encapsulate request update payload data.",
+            "description": "UpdateAssetInput is the input payload to update an asset.",
             "type": "object",
             "properties": {
                 "metadata": {
@@ -2206,7 +2416,7 @@ const docTemplate = `{
             }
         },
         "UpdateLedgerInput": {
-            "description": "UpdateLedgerInput is a struct design to encapsulate request update payload data for ledger.",
+            "description": "UpdateLedgerInput is the input payload to update a ledger.",
             "type": "object",
             "properties": {
                 "metadata": {
@@ -2224,7 +2434,7 @@ const docTemplate = `{
             }
         },
         "UpdateOrganizationInput": {
-            "description": "UpdateOrganizationInput is a struct design to encapsulate request update payload data for organization.",
+            "description": "UpdateOrganizationInput is the input payload to update an organization.",
             "type": "object",
             "required": [
                 "legalName"
@@ -2257,7 +2467,7 @@ const docTemplate = `{
             }
         },
         "UpdatePortfolioInput": {
-            "description": "UpdatePortfolioInput is a struct design to encapsulate payload data.",
+            "description": "UpdatePortfolioInput is the input payload to update a portfolio.",
             "type": "object",
             "properties": {
                 "metadata": {
@@ -2275,7 +2485,7 @@ const docTemplate = `{
             }
         },
         "UpdateProductInput": {
-            "description": "UpdateProductInput is a struct design to encapsulate request update payload data.",
+            "description": "UpdateProductInput is the input payload to update a product.",
             "type": "object",
             "properties": {
                 "metadata": {