+++ components/transaction/internal/bootstrap/http/transaction.go
@@ -30,6 +30,20 @@ type TransactionHandler struct {
 }
 
 // CreateTransactionJSON method that create transaction using JSON
+//
+//	@Summary		Create a Transaction using JSON
+//	@Description	Create a Transaction with the input payload
+//	@Tags			Transactions
+//	@Accept			json
+//	@Produce		json
+//
+//	@Param			organization_id	path		string								true	"Organization ID"
+//	@Param			ledger_id		path		string								true	"Ledger ID"
+//
+//	@Param			transaction		body		transaction.CreateTransactionInput	true	"Transaction Input"
+//	@Param			Midaz-Id		header		string								false	"Request ID"
+//	@Success		200				{object}	transaction.Transaction
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/json [post]
 func (handler *TransactionHandler) CreateTransactionJSON(p any, c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -51,6 +65,20 @@ func (handler *TransactionHandler) CreateTransactionJSON(p any, c *fiber.Ctx) er
 }
 
 // CreateTransactionDSL method that create transaction using DSL
+//
+//	@Summary		Create a Transaction using DSL
+//	@Description	Create a Transaction with the input DSL file
+//	@Tags			Transactions
+//	@Accept			mpfd
+//	@Produce		json
+//
+//	@Param			organization_id	path		string	true	"Organization ID"
+//	@Param			ledger_id		path		string	true	"Ledger ID"
+//
+//	@Param			transaction		formData	file	true	"Transaction DSL file"
+//	@Param			Midaz-Id		header		string	false	"Request ID"
+//	@Success		200				{object}	transaction.Transaction
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/dsl [post]
 func (handler *TransactionHandler) CreateTransactionDSL(c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -99,6 +127,8 @@ func (handler *TransactionHandler) CreateTransactionDSL(c *fiber.Ctx) error {
 }
 
 // CreateTransactionTemplate method that create transaction template
+//
+// TODO: Implement this method and the swagger documentation related to it
 func (handler *TransactionHandler) CreateTransactionTemplate(p any, c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -114,6 +144,8 @@ func (handler *TransactionHandler) CreateTransactionTemplate(p any, c *fiber.Ctx
 }
 
 // CommitTransaction method that commit transaction created before
+//
+// TODO: Implement this method and the swagger documentation related to it
 func (handler *TransactionHandler) CommitTransaction(c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -127,6 +159,8 @@ func (handler *TransactionHandler) CommitTransaction(c *fiber.Ctx) error {
 }
 
 // RevertTransaction method that revert transaction created before
+//
+// TODO: Implement this method and the swagger documentation related to it
 func (handler *TransactionHandler) RevertTransaction(c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -140,6 +174,21 @@ func (handler *TransactionHandler) RevertTransaction(c *fiber.Ctx) error {
 }
 
 // UpdateTransaction method that patch transaction created before
+//
+//	@Summary		Update a Transaction
+//	@Description	Update a Transaction with the input payload
+//	@Tags			Transactions
+//	@Accept			json
+//	@Produce		json
+//
+//	@Param			organization_id	path		string								true	"Organization ID"
+//	@Param			ledger_id		path		string								true	"Ledger ID"
+//	@Param			transaction_id	path		string								true	"Transaction ID"
+//
+//	@Param			transaction		body		transaction.UpdateTransactionInput	true	"Transaction Input"
+//	@Param			Midaz-Id		header		string								false	"Request ID"
+//	@Success		200				{object}	transaction.Transaction
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/{transaction_id} [patch]
 func (handler *TransactionHandler) UpdateTransaction(p any, c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -182,6 +231,19 @@ func (handler *TransactionHandler) UpdateTransaction(p any, c *fiber.Ctx) error 
 }
 
 // GetTransaction method that get transaction created before
+//
+//	@Summary		Get a Transaction by ID
+//	@Description	Get a Transaction with the input ID
+//	@Tags			Transactions
+//	@Produce		json
+//
+//	@Param			organization_id	path		string	true	"Organization ID"
+//	@Param			ledger_id		path		string	true	"Ledger ID"
+//	@Param			transaction_id	path		string	true	"Transaction ID"
+//
+//	@Param			Midaz-Id		header		string	false	"Request ID"
+//	@Success		200				{object}	transaction.Transaction
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/{transaction_id} [get]
 func (handler *TransactionHandler) GetTransaction(c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -209,6 +271,19 @@ func (handler *TransactionHandler) GetTransaction(c *fiber.Ctx) error {
 	return http.OK(c, tran)
 }
 
+// GetAllTransactions method that get all transactions created before
+//
+//	@Summary		Get all Transactions
+//	@Description	Get all Transactions with the input metadata or without metadata
+//	@Tags			Transactions
+//	@Produce		json
+//
+//	@Param			organization_id	path		string	true	"Organization ID"
+//	@Param			ledger_id		path		string	true	"Ledger ID"
+//
+//	@Param			Midaz-Id		header		string	false	"Request ID"
+//	@Success		200				{object}	mpostgres.Pagination{items=[]transaction.Transaction}
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions [get]
 func (handler *TransactionHandler) GetAllTransactions(c *fiber.Ctx) error {
 	ctx := c.UserContext()